Materialize the items when iterating them in Dict's constructor
This fixes a bug that can occur when the items object is changed during iteration.
It can happen when items is sys.modules, coming through inference at some point
and if there's an additional installed library that modifies sys.modules when
it is imported, as it is the case for pytest.AliasModule.

--HG--
branch : 1.4.0