Materialize the items of the dict when building a Dict node from a live dictionary
This fixes a bug that can occur when the dict object is changed during iteration.
It can happen when it is sys.modules, coming through inference at some point
and if there's an additional installed library that modifies sys.modules when
it is imported, as it is the case for pytest.AliasModule.