Enable the warnings in tests through a context manager
Previously, we were enabling blindly all the warnings in
tests files through `warnings.simplefilter`, which had the side effect
of not taking into account the values of the -W flag. This meant
that warnings which should have been silenced, weren't, leading to
a huge output when running the tests with Python 3.5.