Ignore elements of sequences when converting them to AST with const_factory
const_factory is buggy in master, as well as the entire raw_building
module, which means that at given points during rebuilding of ASTs
from live objects, nodes are constructed with Python objects as
elements, leading to various crashes with pylint and with astroid
during certain inference paths. This is fixed in 2.0 though, where
raw_building was completely rewritten, but bringing those changes
back into master will need some more work to be done. Until this
happens, we ignore the elements of sequences passed to const_factory,
since they are not useful in anyway (and the code that uses them
crashes anyway). This is just a temporary solution until we either
merge master with 2.0 or backport raw_building changes from 2.0
to master.