Add a new possible property-descriptor, lazy.
Since we don't understand descriptors at all, we're relying
on a list of possible property-descriptors, that is, decorators
which transforms a method into its value, the exact process that
a property does. While this is a bad approach, it works, at least
until we can provide support for this high level of understanding.
Closes #279