Support Jython on Travis
This is a bit complicated for a couple of reasons. Jython is not
available out of the box on Travis and needs to be installed manually,
but the released version (2.7.0) doesn't play nice with pip, resulting
in being completely unusable. Thus we're building Jython ourselves
from the repository, which works, as long as tox can find where the
jython executable is.
Closes #288