Remove do_import_module and real_name from ImportFrom and Import nodes.
do_import_module was broken for Import, since it was relying on a
missing attribute `modname`. do_import_module
in fact required all the time for a name parameter to be
passed as an argument. The inherent problem is that
Import doesn't have an underlying module name, but more of them,
in the .names attribute. Thus requiring for do_import_module to
pick one name from .names or to return multiple modules was
deemed inappropiate and the method became a function which
always requires the import name to be given.
do_import_module and real_name are now functions in interpreter.util.
Closes #293