Add specific checking for features instead of catching AttributeError in inference
Certain blocks were protected with an AttributeError handler in order to
ignore certain nodes which didn't provide an expected feature, such
as .getattr and .getitem. This was replaced with an attribute check
instead, in order to limit the scope of the exception handling, which
could have concealed other bugs. We could check, in the future,
for protocol classes instead, but in short term, this is the best
approach. Closes #294