Fix minor issues in packaging and testing files.
I removed "recursive-include brain/" in MANIFEST.in because it doesn't
do anything, probably because Python files in subdirectories of
packages are automatically included.  For some reason "graft
astroid/tests/testdata" doesn't work, maybe because tests/ contains an
`__init__.py`.  I also altered coverage to include only astroid files
by using --source in coverage run call, rather than --ignore in the
reports.  Note that when tox is called from the root of the repository
at the moment, astroid is imported from the files in the *repository*,
not the installed files in the virtualenvs in .tox/.  Meanwhile, the
tests themselves are run from the virtualenvs in .tox/, which accounts
for why coverage shows 0% for all the unittest files in tests/.