Try to see if the object from the determined module is actually the object we're building
They can be different when in the original module (let's say an __init__ file) the object
is imported under a name and then used as a base class for a class with the same name as
the object in question, paired with the fact that the initial object is faking its name
inside is original module. This happens for instance with the bitarray module, which has
a bitarray._bitarray C extensions, which defines a bitarray._bitarray._bitarray type,
the latter faking its name as bitarray._bitarray. In the __init__ file, the original
type gets used as a base class in the actual bitarray class that the module exports.
What happens is that this behaviour will introduce a cycle in the form of an ImportFrom
as in 'from bitarray._bitarray import _bitarray as _bitarray` when analyzing the bitarray._bitarray
module, which means that the resulting ast will reimport itself, leading to not inferring
anything from the module. The solution we're using is to check that the module where we expected
to have our object actually has the object we want.
Unfortunately I couldn't find a Python test through which we can reflect this behaviour