Apply the strings / bytes patching only when the builtins module gets built
This patch fixes a problem with the transforms on Python 2, which were never called due
to a bad interaction between a transform and the underlying inference mechanism.
The string / byte class was patched before the brain modules were loaded, since
extend_str was called at the module level. Inside extend_str we were using .locals,
which triggers .delayed_assignments, which can trigger as well inference. In the case
of Python 2, this led to importing of the six module before actually having the transforms
ready for it, which means that no transforming was ever done for some of these modules.
By delaying the patching of the builtins module until builtins is ready, we alleviate
this problem.