Don't restrict building of builtin type's member if they are already in special_attributes
The introduction of the special attribute models led to an interesting bug, where
some of the members of the builtin type were not added into it, since they were
already defined at the level of the special_attributes. The latter is just a declaration
of what attributes a type can have, usually leading to introspecting the type itself for
retrieving the value of the special attribute.