Try to patch special_attributes instead of returning a new node
It seems we have some problems with transforms that are creating new nodes.
While they work okay in most of the cases, we found a problem with regard
to pylint. For instance, if we recreate a function and return it
from a transform, pylint can see that we had two instances of the same
function in the module locals and will emit an error saying that the
function is redefined. In fact, the old node is still available
in the module locals, while it should have been replaced.
Unfortunately, there is not much we can do right now, instead of having
something as the zipper integrated.
This patch would fix the undefined-variable false positives we had in pylint.