Use nodes.Unknown instead of Uninferable for specifying unknown inference results
In the context of binop inference for lists and tuples, we were rebuilding
a collection with the inferred elements. We were using Uninferable for specifying inference failures, but this leads to having non-AST nodes in collection's elts attribute, which should not happen. Instead, we replaced Uninferable with Unknown, meaning now we have an AST object.