Don't raise DuplicateBaseError when classes at different locations are used
For instance, one can implement a namedtuple base class, which gets reused
on a class with the same name later on in the file. Until now, we considered
these two classes as being the same, because they shared the name, but in fact
they are different, being created at different locations and through different
means.

Close PyCQA/pylint#1458