Remove flaky tests for random.sample
Recently we added support for inferring random.sample return values. But in the cases
where there was something amiss with the function call (too many arguments, too few,
wrong type), we were falling back on the default inference, as it is the normal approach
for inference tips. The problem is that the inference is inconsistent depending on the
Python version. On Python 3.6 it correctly infers a list, but with the elements being None,
while on older versions it returns Uninferable. Since we want to just test that
our brain tip works correctly, we can just remove the other tests and let only
the brain tip test to live.