Fix import of symlinks outside of path
Makes load_module_from_file able to import files within the path that
are just a symlink to a file outside of the path.
modutils is using realpath at the moment which will resolve the symlink
and don't allow the import as the file is detected outside of the
available paths. By checking both abspath and realpath this patches add
support for such scenarios.