Remove warnings for nodes that were replaced multiple time
It is possible to have nodes that are transformed multiple times,
so this warning doesn't make that much sense. It is also a bit useless
in case we replace the node even once, since it only verifies that the returned
node from a transform is the same node that was passed in.