Include abspath of sys.path on searches at load_from_module
At the moment `load_from_module` is resolving all paths in `sys.path`
to validate that the module being import is in a folder part of the
path. This is done to resolve symlinks.

This can be an issue though it both the file being imported and the path
in `sys.path` resolve to different locations. Even if the symlinks are
importable, the resolution of the paths in `sys.path` might not.