Don't allow max_inferable to be set by an environment variable.
This would introduce a new method of configuring astroid & pylint, while the
usual approach we used so far was to set a flag in the MANAGER itself it we'd
need to customize some behaviour (extension_package_whitelist and friends).

Also renamed the flag to `max_inferable_values` to be more suggestive on what it does.