Remove the restore_path() method
This was initially added in e642ba33ba1bdde04ac9f0c75a25dc40131c55e7 so astroid won't
skip undesired inference branches. The method though doesn't seem to influence anything
as the added test is now passing regardless if we restore the path or not. The correct
approach would have been to clone the context instead. But given that the context is usually
used for caching already inferenced objects, we don't necessarily need to clone it here, but rather
in the low level inference, where we'd know that the inference path will diverge.