Cache the result of get_assign_nodes
The result is usually static, but we're hitting this function quite
a lot for big projects such as pandas. This change is a matter of compromising
between the amount of memory astroid needs for analysis (e.g. storing the
results in memory) vs the wasted CPU time in calling the same
function over and over.