Revert path copying in InferenceContext()
While copying the path solves some inference issues, it also leads to
horrenduous performance hits on large libraries such as pandas.
We still want to have the path copying, but we need a better understanding
on why that is needed, as well as a change on how we operate with the inference
contexts.