Make the astroid bootstrapping take care of builtins such as Generator
Currently the astroid bootstrapping function, which builds the builtins
module and relevant builtin objects, does not rebuild objects such as
generators. This is causing a problem when the cache is cleared, e.g.
from pylint's side, given that objects such as generator will not be
present any longer in the builtins module created *after* the cache was
cleared with MANAGER.cache_clear(). This behaviour was making that function
useless from this point of view.
The new change instead makes sure to build everything in the same
function.

Close #446