Drop a superfluous and wrong callcontext when inferring the result of a context manager
The callcontext was containing the instance of the class (self) as the sole argument,
but this was tripping the inference of unknown arguments coming down into an
instance, such as the one from the commit's example.
By dropping the superfluous callcontext, the inference can no longer assume
that the first argument is the instance of the class, leading to wrongly
infer `self.client` as the first argument of the instantiation call.

Close PyCQA/pylint#2859