Set the line number of decorated functions for python 3.8
Python 3.8 sets the line number of a decorated function
to be the actual line number of the function, but the
previous versions expected the decorator's line number instead.
We reset the function's line number to that of the
first decorator to maintain backward compatibility.
It's not ideal but this discrepancy was baked into
the framework for *years*.