Remove `__class__` from FunctionDef locals, instead look it up in the current scope
Previously `__class__` was added as an implicit local for methods, but that causes some issues
as it now is counted as a local instead of a special object created by the compiler.
A better way to be able to find `__class__` is to look it up in the current scope.

Close PyCQA/pylint#1825