Avoid a call to `getcwd` if it is not needed
When profiling locally, approximately 7% of a full pylint run is spent in `os.getcwd`, despite its value effectively not being used if `context_file` is `None`.  A quick grep over this project and others seem to indicate that `context_file` is rarely used, so this should almost entirely get rid of this system call.