Add the float_power function in the brain and in its associated test
The bug PyCQA/pylint#3319 was caused due to a missing function, named float_power inside the brain_numpy_core_umath module. This brain is used to infer all the numpy's ufunc functions. The problem is that in the website documentation of numpy, in the section that list all those functions, there is not entry dealing with the float_power function.

Also adds 5 missing functions detected by @texadactyl

Closes PyCQA/pylint#3319