Add a call to register_transform for each numpy function in case the current node is an astroid.Name instance
The problem was that astroid could not infer the result of a call to `numpy.append`
because this function calls the `concatenate` function.
This last function is inferred thanks to the `brain_numpy_core_multiarray` module
but only when the corresponding node is an `astroid.Attribute` (for example numpy.concatenate). 
It turns out that in the source of the append function the node that realises the call to concatenate is a `astroid.Name`. Thus the correction proposed here is to register
the concatenate inference tip function in order to apply it, also, to `astroid.Name`.

Close #666