numpy.astype now returns a ndarray object
This fixes PyCQA/pylint#3332. The bug was due to the fact that datetime64 class
inherited from generic one and in this last class the astype method was returning
uninferable. Now it returns ndarray object.

This fix should be generalized to all other methods of the generic class as the numpy doc specifies that the generic object API should be the same as the ndarray one.