Raise ``AttributeInferenceError`` when ``getattr()`` receives an empty name
If `Module.getattr` received an empty string (as a result of inference for example),
`astroid` would have returned the same Module again, which leads to false positives
in pylint, since the expected output was of a different type.

Rather than allowing empty names to pass through `getattr()`, we simply raise
an error earlier.

Close PyCQA/pylint#2991