Add an optional `path` parameter to `modpath_from_file`
Because we'll no longer have the current working directory, nor the module's directory
in `sys.path`, we need to be able to pass additional paths to be checked against when
trying to load a module. Most of the functions in this module already accepted
a `path` parameter, except `modpath_from_file`.