Merge TreeRebuilder3 into TreeRebuilder
These were separated to allow us to add support for parsing Python 2 files
as well as Python 3. Unfortunately that is no longer in our plans and since
the code was already broken (e.g. the Raise node was receiving unexpected arguments),
it's better to rip the bandaid and merge the two classes together.