Allow slots added dynamically to a class to still be inferred
In 2aa27e9aed6ffcba4a61655e291e852ecd001549 `ClassDef.igetattr`
was modified to only grab the first item from the result of `getattr`,
in order to avoid looking up attributes in the ancestors path when
inferring attributes for a given class. This had the side effect
that we'd omit attribute definitions happening in the same scope,
such as augmented assignments, which in turn might have affected
other capabilities, such as slots inference.

This commit changes the approach a bit and keeps all attributes
as long as all of them are from the same class (be it current
or an ancestor)

Close PyCQA/pylint#2334