Don't cache exception values in the module cache
Exceptions hold tracebacks, which can get very large (especially for
highly recursive functions). Especially if the exceptions are being used
for control flow instead of actual exception management, we should be
throwing them away as fast as possible.

This commit deals with this with a bit of a hacky solution, but really
the way the module cache handles things ought to be reworked a bit.