Fix inference of instance attributes defined in base class
Ref #932. The cause of this appears to be failing to specify the
lookupname on the inference context in BaseInstance.igetattr.
Additionally, the context path key was not set in a manner that matched
other usages. This appears to date back to at least 2015 when context
path keys were changed. The usage as a tuple instead of a node with
implict lookupname was changed last in 3d342e85 but the lookup key was
preserved throughout various changes since then.