Fix test definition of igetattr recursion and context_manager_inference
Ref #663. This test did not actually check for regression of the issue
fixed in 55076ca0 (i.e. it also passed on c87bea17 before the fix was
applied). Additionally, it over-specified the behaviour it was
attempting to check: whether the value returned from the context manager
was Uninferable was not directly relevant to the test, so when this
value changed due to unrelated fixes in inference, this test failed.