Fix strong references to mutable objects in InferenceContext
Ref #926. Since InferenceContext.clone passes self.path when
constructing a new instance, these two objects will hold references to
the same underlying set object. This caused modifications to
context.path further down in the stack to be reflected in "earlier"
objects and this would incorrectly detect a loop if the same node was
used twice in the inference of a some node (instead of only if an
inference result was used in inferring itself). As an illustration:

[stmt1 = something.one, stmt1 = something.two]
-> infer stmt1       <-+
-> -> infer something -+  writes back to parent context.path -+
-> <- <returns>                                               |
-> infer stmt2           affecting this context.path        <-+
-> -> infer something -> sees something in context.path,
                         decorators.path_wrapper returns None